backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
redis==5.0.4
PyMySQL==1.1.1
python-dotenv==1.0.1
